projects redirect and update, remove directory and subdirectory, updated people query
