+++ templates/views/projects/project.hbs
@@ -92,7 +92,7 @@
 									    <div class="glide__wrapper">
 									        
 									        <ul class="glide__track" style="margin-left:-40px">
-														{{# each projectImageObjects}}
+														{{#each projectImageObjects}}
 															<li class="glide__slide">
 																<img src="{{{cloudinaryUrl img width=500 quality=80}}}" alt="{{@key}}">
 																<div class="caption lead">{{caption}}</div>
