+++ templates/views/projects/projectsAll.hbs
